!4. refactoring: объеденил PlatformDto. переписал PlatformMapper
