+++ src/main/java/ru/hhschool/segment/mapper/PlatformMapper.java
@@ -1,30 +1,5 @@
 package ru.hhschool.segment.mapper;
 
-import java.util.List;
-import ru.hhschool.segment.model.dto.platform.PlatformDto;
-import ru.hhschool.segment.model.entity.Platform;
-
-public class PlatformMapper {
-  public static PlatformDto platformToDto(Platform platform) {
-    PlatformDto platformDto = new PlatformDto(
-        platform.getId(),
-        platform.getPlatform().toString(),
-        platform.getApplicationVersion()
-    );
-    return platformDto;
-  }
-
-  public static List<PlatformDto> platformListToDtoList(List<Platform> platformList) {
-    if (platformList == null) {
-      return List.of();
-    }
-    return platformList.stream()
-        .map(PlatformMapper::platformToDto)
-        .toList();
-  }
-}
-package ru.hhschool.segment.mapper;
-
 import java.util.Comparator;
 import java.util.List;
 import ru.hhschool.segment.model.dto.PlatformDto;
@@ -32,17 +7,18 @@ import ru.hhschool.segment.model.entity.Platform;
 
 public class PlatformMapper {
 
-  private static PlatformDto toSegmentViewPlatformDto(Platform platform) {
-    PlatformDto segmentViewApplicationDto = new PlatformDto(
+  private static PlatformDto toDto(Platform platform) {
+    PlatformDto platformDto = new PlatformDto(
+        platform.getId(),
         platform.getPlatform(),
         platform.getApplicationVersion()
     );
-    return segmentViewApplicationDto;
+    return platformDto;
   }
 
-  public static List<PlatformDto> toDtoForSelectedSegmentViewPage(List<Platform> platforms) {
+  public static List<PlatformDto> toDtoList(List<Platform> platforms) {
     return platforms.stream()
-        .map(platform -> toSegmentViewPlatformDto(platform))
+        .map(platform -> toDto(platform))
         .sorted(Comparator.comparing(PlatformDto::getPlatform))
         .toList();
   }
