+++ src/main/java/ru/hhschool/segment/mapper/screen/ScreenMapper.java
@@ -1,7 +1,7 @@
 package ru.hhschool.segment.mapper.screen;
 
 import java.util.List;
-import ru.hhschool.segment.model.dto.platform.PlatformDto;
+import ru.hhschool.segment.model.dto.PlatformDto;
 import ru.hhschool.segment.model.dto.screen.ScreenCreateDto;
 import ru.hhschool.segment.model.dto.screen.ScreenDto;
 import ru.hhschool.segment.model.entity.Question;
